+++ priv/repo/migrations/20180730150202_add_guild_ban_add_to_modlog_event_enum.exs
@@ -1,0 +1,8 @@
+defmodule Bolt.Repo.Migrations.AddGuildBanAddToModlogEventEnum do
+  use Ecto.Migration
+  @disable_ddl_transaction true
+
+  def up do
+    execute("ALTER TYPE modlog_event ADD VALUE 'GUILD_BAN_ADD';")
+  end
+end