+++ priv/repo/migrations/20180730150302_add_guild_ban_remove_to_modlog_event_enum.exs
@@ -1,0 +1,8 @@
+defmodule Bolt.Repo.Migrations.AddGuildBanRemoveToModlogEventEnum do
+  use Ecto.Migration
+  @disable_ddl_transaction true
+
+  def up do
+    execute("ALTER TYPE modlog_event ADD VALUE 'GUILD_BAN_REMOVE';")
+  end
+end