+++ lib/bolt/modlog.ex
@@ -16,6 +16,8 @@ defmodule Bolt.ModLog do
     "INFRACTION_UPDATE" => "🖋",
     "INFRACTION_EVENTS" => "⏲",
     "CHANNEL_CREATE" => "#⃣",
+    "GUILD_BAN_ADD" => "🔨",
+    "GUILD_BAN_REMOVE" => "🤝",
     "GUILD_MEMBER_ADD" => "📥",
     "GUILD_MEMBER_REMOVE" => "📤",
     "GUILD_MEMBER_UPDATE" => "♻",