+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- add the `GUILD_BAN_ADD` and `GUILD_BAN_REMOVE` modlog events
 - show username and discrim in addition to mention in `inrole` command
 - update to newest Elixir version
 - renamed USW "filters" to "rules"